Added Edit

+++ src/main/java/switer/repos/MessageRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 import switer.domains.Message;
 
 import java.util.List;
-public interface MessageRepository extends CrudRepository<Message, Integer>
+public interface MessageRepository extends CrudRepository<Message, Long>
 {
     List<Message> findByTag(String tag);
 }
